Hide the gene validation checkmarks to give the text more space
There's feedback for invalid entries with the red underline.  The underline result is perceptively instant for reasonably-sized gene lists.  For feedback for large gene lists, the text fades out while validating.  Given that the checkbox does not give any information not visible elsewhere, this change makes better use of space.